requirements.txt
appdirs==1.4.4
asgiref==3.2.10
aspy.refactor-imports==2.1.1
attrs==20.2.0
black==20.8b1
cached-property==1.5.2
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
Django==3.1.2
django-environ==0.4.5
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
eradicate==1.0
filelock==3.0.12
flake8==3.8.4
flake8-bugbear==20.1.4
flake8-builtins==1.5.3
flake8-comprehensions==3.2.3
flake8-docstrings==1.5.0
flake8-eradicate==0.4.0
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-pep3101==1.3.0
flake8-plugin-utils==1.3.1
flake8-polyfill==1.0.2
flake8-print==3.1.4
flake8-printf-formatting==1.1.0
flake8-pytest-style==1.3.0
identify==1.5.6
importlib-metadata==2.0.0
isort==5.6.2
marshmallow==3.8.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.5.0
pathspec==0.8.0
pep8-naming==0.11.1
pkg-resources==0.0.0
pre-commit==2.7.1
psycopg2==2.8.6
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.9.27
seed-isort-config==2.2.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
virtualenv==20.0.33
zipp==3.3.0
